[Refactor] assets 경로 수정
