+++ src/components/breadDetail/DetailModal.jsx
@@ -1,7 +1,8 @@
 import styled, { css } from 'styled-components';
 import { useRecoilValue, useSetRecoilState } from 'recoil';
 import { modalState, readingDataList, idState } from '../../atoms/fishBreadList';
-import '../../../public/assets/font/font703.css';
+//import '../../../public/assets/font/font703.css';
+import './assets/font/font703.css';
 import { useEffect, useState } from 'react';
 
 const TypeObj = {
@@ -92,8 +93,8 @@ const ModalContainer = styled.div`
   align-items: center;
 
   ${({ backSrc }) => css`
-    background: #fff url('../../../public/assets/images/breadDetail/${backSrc}.png') no-repeat
-      center/cover;
+    //background: #fff url('../../../public/assets/images/breadDetail/${backSrc}.png') no-repeat center/cover;
+    background: #fff url('./assets/images/breadDetail/${backSrc}.png') no-repeat center/cover;
   `}
 `;
 
@@ -119,8 +120,8 @@ const MessageWrapper = styled.div`
   margin-top: 30px;
 
   ${({ letterSrc }) => css`
-    background: url('../../../public/assets/images/breadDetail/${letterSrc}.png') no-repeat center /
-      contain;
+    //background: url('../../../public/assets/images/breadDetail/${letterSrc}.png') no-repeat center / contain;
+    background: url('./assets/images/breadDetail/${letterSrc}.png') no-repeat center / contain;
   `}
 `;
