+++ src/components/breadDetail/list/DetailList.jsx
@@ -9,7 +9,6 @@ import { dataList, tapIndexState } from '../../../atoms/fishBreadList';
 import { getBreadListData } from '../../../utils/fetchBreadDetail';
 import { Cookies } from 'react-cookie';
 import styled from 'styled-components';
-import mailbox from '../../../../public/assets/images/breadDetail/mailbox.png';
 
 const BREAD_DATA = {
   content: [
@@ -200,16 +199,16 @@ const DetailLists = styled.div`
   max-width: 354px;
   max-height: 551px;
   position: relative;
-  background: url('../../../../public/assets/images/breadDetail/mailbox.png') no-repeat
-    center/contain;
+  //background: url('../../../../public/assets/images/breadDetail/mailbox.png') no-repeat center/contain;
+  background: url('./assets/images/breadDetail/mailbox.png') no-repeat center/contain;
 `;
 
 const TurnBack = styled.div`
   text-indent: -9999px;
   width: 50px;
   height: 50px;
-  background: url('../../../../public/assets/images/breadDetail/turnBack.png') no-repeat
-    center/cover;
+  //background: url('../../../../public/assets/images/breadDetail/turnBack.png') no-repeat center/cover;
+  background: url('./assets/images/breadDetail/turnBack.png') no-repeat center/cover;
   cursor: pointer;
   margin-bottom: 14px;
 `;
