+++ src/components/breadDetail/list/DetailListItem.jsx
@@ -1,6 +1,7 @@
 import { useEffect, useState } from 'react';
 import styled from 'styled-components';
-import '../../../../public/assets/font/font703.css';
+//import '../../../../public/assets/font/font703.css';
+import './public/assets/font/font703.css';
 
 const TypeObj = {
   밀가루: '1',
@@ -39,7 +40,8 @@ function DetailListItem({ data, onClickBread }) {
     <ItemWrapper>
       <ItemNickname>{senderNickname}</ItemNickname>
       <ItemImage
-        src={`../../../../public/assets/images/breadDetail/${breadType}.png`}
+        //src={`../../../../public/assets/images/breadDetail/${breadType}.png`}
+        src={`./assets/images/breadDetail/${breadType}.png`}
         onClick={() => onClickBread(id)}
       />
     </ItemWrapper>
