+++ src/main.jsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
-import '../public/assets/font/font.css';
+//import '../public/assets/font/font.css';
+import './assets/font/font.css';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
